Update runme.sh to work with submodules
This is an initial implementation of runme.sh so
it can work with submodules rather than direct
checkouts.  It also introduces a variable option
BOOTLOADER_ONLY which will bypass building the images
and kernel and only build the packages required to
boot the machine.

Note currently only BOOTLOADER_ONLY is functional in
this branch.  I will work on merging additional patches
to test the other features of runme.sh moving forward.

Fixes: edk2 is now bootable.

Signed-off-by: Jon Nettleton <jon@solid-run.com>