Unify boot images of SPI, SD and eMMC and automate RCW creation
Using conditional load and jump in RCW and boot=auto as a parameter for
ATF it is possible now to build a single image that can be booted from
SPI, SD or eMMC.

Still there is a small difference BL2 stage where on SPI it is located
on address 0x0 and SD/eMMC is located with 0x1000 (8 blocks of 512
bytes) offset.

Additional feature is runtime RCW file creation; this will be further
enhanced to create the SERDES and PLL multipliers creation in runtime
too.

Signed-off-by: Rabeeh Khoury <rabeeh@solid-run.com>