Half-twins motherboard: fix dpmac.16 to dpmac.18
The issue is that dpmac.16 to dpmac.18 links up, but unable to
communicate thru those ports.
This patch series does -
1. Fix dpmac.17 phy-handle type since by default the dpmac is RGMII, and
in half-twins case it is thru SGMII.
2. Add to the build script the usage of MC firmware 10.28.100; which is
based on 10.28.1 that fixes the tx/rx thru dpmac.16, dpmac.17 and
dpmac.18. This firmware is only for LSDK-21.08 and will be fixed in
later LSDK releases with newer MC firmwares.
3. Limit recycle queues in the half twins DPC to 1G; this makes it
possible for dpmac.18 to be able to transmit

Signed-off-by: Rabeeh Khoury <rabeeh@solid-run.com>