half-twins: connect sfp cages link status to LEDs
Directly connect the sfp drivers's (sfp.c) link up and down events to an
LED node in device-tree. This solution is a workaround to the
shortcomings of the generic led triggers framework in combination with
multiple dynamic network interfaces, hot-pluggable phys (inside sfp
modules) and the dpaa2 driver in particular.

Signed-off-by: Josua Mayer <josua@solid-run.com>