linux: add support for link-status LEDs to dpmac objects
To allow reliable link-status indication when exporting native
interfaces to dpdk, add support for controlling an LED to dpmac driver
link up- and down- functions.

Note that this is a hack and not a proper solution - with its own
particular shortcomings. Most notable is the lack of status indication
while Linux is controlling the interface through the dpaa2 ethernet
driver!

Signed-off-by: Josua Mayer <josua@solid-run.com>