runme: fix nameserver within chroot
Change the ip configuration for the temporary VM to be performed
directly by the kernel; then copy the kernel reported resolv config to
rootfs + chroot.

This will avoid firewall issues that may be experienced in paranoid
homes, or coprorate networks where direct access to public DNS servers
is not allowed.